switch to 2nd widget after first interval
*Problem*

When using the widget switcher it displays the first element twice, then iterates over the rest consistently. To be able to synchronise different widgets with intervals of a multiple of each other, they become out of sync straight await.

*Solution*

Set the current Pos to 1 at the beginning, rather than 0. This will switch to 1 after the first interval.